% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_base.R
\name{arrange_base}
\alias{arrange_base}
\title{Arrange or Reorder Rows of a Data Frame by Columns}
\usage{
arrange_base(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Column names to determine the sort order.}
}
\value{
A sorted data frame.
}
\description{
This function sorts a data frame based on the provided columns.
}
\keyword{internal}
