% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_roof.R
\name{change_polygon_bevel}
\alias{change_polygon_bevel}
\title{Change an existing polygon bevel's bevel profile.}
\usage{
change_polygon_bevel(
  skeleton_polygons,
  bevel_offsets = NULL,
  bevel_heights = NULL,
  set_max_height = FALSE,
  max_height = 1,
  offset = 0,
  base = FALSE,
  raw_offsets = FALSE,
  raw_heights = FALSE,
  swap_yz = TRUE,
  verbose = TRUE,
  double_sided = FALSE
)
}
\arguments{
\item{skeleton_polygons}{Default `NULL`. A straight skeleton generated from the `generate_beveled_polygon` function when
`return_skeleton_polygons = TRUE`.}

\item{bevel_offsets}{Default `NULL`. The offset(s) of the bevel.}

\item{bevel_heights}{Default is set to `bevel_offsets`. Numeric vector specifying the heights of the bevels. Must be of the same length as `bevel_offsets`.}

\item{set_max_height}{Default `FALSE`. A logical flag that controls whether to set the max height of the polygon based on the `max_height` argument.}

\item{offset}{Default `0`. The vertical offset of the polygon.}

\item{base}{Default `FALSE`. A logical flag that controls whether to generate the bottom of the polygon.}

\item{swap_yz}{Default `TRUE`. A logical flag that controls whether to swap the y and z coordinates in the resulting mesh.
If `TRUE`, the y and z coordinates will be swapped.}
}
\value{
A 3D mesh of the beveled polygon model.
}
\description{
This function generates a beveled 3D polygon model from a straight skeleton.
}
