% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ss_to_polygons.R
\name{convert_ss_to_polygons}
\alias{convert_ss_to_polygons}
\title{Get Face IDs}
\usage{
convert_ss_to_polygons(ss, progress = FALSE)
}
\arguments{
\item{ss}{}
}
\description{
Get Face IDs
}
\keyword{internal}
