% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-functions.R
\name{determinant2x2}
\alias{determinant2x2}
\title{Title}
\usage{
determinant2x2(v1, v2)
}
\arguments{
\item{v1}{Vector 1}

\item{v2}{Vector 2}
}
\value{
Value
}
\description{
Title
}
\examples{
#None
}
\keyword{internal}
