% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_nodes_within_tolerance.R
\name{discretize_and_merge_nodes}
\alias{discretize_and_merge_nodes}
\title{Discretize and Merge Nodes}
\usage{
discretize_and_merge_nodes(skeleton, tolerance = 1e-05)
}
\arguments{
\item{ss}{}
}
\description{
Discretize and Merge Nodes
}
\keyword{internal}
