% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrude_polygon.R
\name{extrude_sides}
\alias{extrude_sides}
\title{Extrude Sides of Polygon}
\usage{
extrude_sides(vertices, holes = list(), bottom = 0, top = 1)
}
\description{
Extrude Sides of Polygon
}
\keyword{internal}
