% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dangling_nodes.R
\name{find_dangling_nodes}
\alias{find_dangling_nodes}
\title{Discretize and Merge Nodes}
\usage{
find_dangling_nodes(skeleton)
}
\arguments{
\item{ss}{}
}
\description{
Discretize and Merge Nodes
}
\keyword{internal}
