% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_bevel.R
\name{generate_bevel}
\alias{generate_bevel}
\title{Generate 2D Bevel Profile for 3D Polygons}
\usage{
generate_bevel(
  bevel_type = "angled",
  percent_coverage = 0.1,
  max_height = 0.1,
  angle = 45,
  curve_points = 50,
  reverse = FALSE,
  initial_height = 0
)
}
\arguments{
\item{bevel_type}{Character `NULL`. Type of the bevel, one of the following options:
- "circular": Creates a rounded bevel resembling a quarter-circle.
- "exp": Creates an exponential curve, starting slow and accelerating.
- "bump": Creates a bump-like profile, rising and falling within the coverage.
- "step": Generates a step-like bevel with a flat top.
- "block": Generates a block-like bevel, jumping straight to the max_height and back to the base.
- "angled": Generates a straight angled bevel. You can optionally set the 'angle' parameter for this bevel.}

\item{percent_coverage}{Numeric `0.1`. Percentage of the curve to be covered by the bevel.}

\item{max_height}{Numeric `0.1`. The maximum height of the bevel, as a percentage of the final height.}

\item{angle}{Numeric `NULL`. Optional angle parameter in degrees for angular bevels.}

\item{curve_points}{Numeric `50`. Number of points to plot for curve-based bevels.}

\item{reverse}{Logical `FALSE`. If TRUE, the curve is reversed.}

\item{initial_height}{Numeric `0`. The initial height from which the bevel starts. The bevel is rescaled to fit within the range from initial_height to max_height.}
}
\value{
List containing 'x' and 'y', which are the coordinates of the 2D bevel profile.
}
\description{
Generate 2D Bevel Profile for 3D Polygons
}
\examples{
#Generate a single bevel profile and plot it
coords = generate_bevel("circular", 0.2, 0.2)
plot(coords$x, coords$y, type = 'l', axes = FALSE, mar = c(0, 0, 0, 0))
title(main = "Bevel Type: Circular")
box()

# Plot all bevel profiles in a grid
par(mfrow = c(4, 3),mai = c(0.2, 0.2, 0.5, 0.2))
types = rep(c("circular", "exp", "bump", "step", "block", "angled"), 2)
rev = c(rep(FALSE, 6), rep(TRUE, 6))
for(i in seq_len(length(types))) {
  coords = generate_bevel(types[i], 0.5, 0.5, 45, reverse = rev[i])
  plot(coords$x, coords$y, type = 'l', main = paste("Bevel Type:", types[i]), axes = FALSE,
       xlim = c(-0.1, 1.1), ylim = c(-0.2, 0.8), mar = c(0, 0, 0, 0),xlab = "", ylab = "")
  box()
}
}
