% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_roof.R
\name{generate_beveled_polygon}
\alias{generate_beveled_polygon}
\title{Generate a beveled 3D polygon}
\usage{
generate_beveled_polygon(
  skeleton,
  bevel_offsets,
  bevel_heights = NULL,
  set_max_height = FALSE,
  max_height = 1,
  offset = 0,
  base = FALSE,
  raw_offsets = FALSE,
  raw_heights = FALSE,
  swap_yz = TRUE,
  verbose = TRUE,
  double_sided = FALSE,
  return_skeleton_polygons = FALSE
)
}
\arguments{
\item{skeleton}{Default `NULL`. A straight skeleton generated from the `skeletonize` function.}

\item{bevel_offsets}{Default `NULL`. The offset(s) of the bevel.}

\item{bevel_heights}{Default is set to `bevel_offsets`. Numeric vector specifying the heights of the bevels. Must be of the same length as `bevel_offsets`.}

\item{set_max_height}{Default `FALSE`. A logical flag that controls whether to set the max height of the roof based on the `max_height` argument.}

\item{max_height}{Default `1`. The maximum height of the polygon.}

\item{offset}{Default `0`. The vertical offset of the polygon.}

\item{base}{Default `FALSE`. A logical flag that controls whether to generate the bottom of the polygon.}

\item{swap_yz}{Default `TRUE`. A logical flag that controls whether to swap the y and z coordinates in the resulting mesh.
If `TRUE`, the y and z coordinates will be swapped.}

\item{verbose}{Default `TRUE`. A logical flag to control whether a progress bar is displayed during roof generation.}
}
\value{
A 3D mesh of the beveled polygon model.
}
\description{
This function generates a beveled 3D polygon from a straight skeleton.
}
