% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_roof.R
\name{generate_new_polygon_heights}
\alias{generate_new_polygon_heights}
\title{Generate a beveled 3D roof model}
\usage{
generate_new_polygon_heights(
  skeleton_polygons,
  bevel_offsets = NULL,
  bevel_heights = NULL,
  set_max_height = FALSE,
  max_height = 1,
  offset_roof = 0,
  base = FALSE,
  raw_offsets = FALSE,
  raw_heights = FALSE,
  swap_yz = TRUE,
  verbose = TRUE,
  double_sided = FALSE
)
}
\arguments{
\item{skeleton_polygons}{Default `NULL`. A straight skeleton generated from the `generate_roof_beveled` function when
`return_skeleton_polygons = TRUE`.}

\item{bevel_offsets}{Default `NULL`. The offset(s) of the bevel.}

\item{bevel_heights}{Default is set to `bevel_offsets`. Numeric vector specifying the heights of the bevels. Must be of the same length as `bevel_offsets`.}

\item{set_max_height}{Default `FALSE`. A logical flag that controls whether to set the max height of the roof based on the `max_height` argument.}

\item{offset_roof}{Default `0`. The offset of the roof.}

\item{base}{Default `FALSE`. A logical flag that controls whether to generate the bottom of the roof.}

\item{swap_yz}{Default `TRUE`. A logical flag that controls whether to swap the y and z coordinates in the resulting mesh.
If `TRUE`, the y and z coordinates will be swapped.}
}
\value{
A 3D mesh of the beveled roof model.
}
\description{
This function generates a beveled 3D roof model from a straight skeleton.
}
