% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_offset_links_nodes.R
\name{generate_offset_links_nodes}
\alias{generate_offset_links_nodes}
\title{Generate Offset Polygon}
\usage{
generate_offset_links_nodes(
  ss,
  offsets,
  return_polys = FALSE,
  progress = FALSE
)
}
\description{
Generate Offset Polygon
}
\keyword{internal}
