% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_offset_polygon.R
\name{generate_offset_polygon}
\alias{generate_offset_polygon}
\title{Generate an offset polygon}
\usage{
generate_offset_polygon(ss, offset)
}
\arguments{
\item{ss}{Default `NULL`. A straight skeleton generated from the `skeletonize` function.}

\item{offset}{Default `NULL`. The offset of the polygon.}
}
\value{
A list of data frames, each representing a polygon offset by the specified amount.
}
\description{
This function generates an offset polygon from a straight skeleton.
}
