% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_roof.R
\name{generate_roof}
\alias{generate_roof}
\title{Generate a 3D roof model}
\usage{
generate_roof(
  skeleton,
  max_height = NA,
  offset_roof = 0,
  bottom = FALSE,
  swap_yz = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{skeleton}{Default `NULL`. A straight skeleton generated from the `skeletonize` function.}

\item{max_height}{Default `NA`. The maximum height of the roof.}

\item{offset_roof}{Default `0`. The offset of the roof.}

\item{bottom}{Default `FALSE`. A logical flag that controls whether to generate the bottom of the roof.}

\item{swap_yz}{Default `TRUE`. A logical flag that controls whether to swap the y and z coordinates in the resulting mesh.
If `TRUE`, the y and z coordinates will be swapped.}

\item{verbose}{Default `TRUE`. A logical flag to control whether a progress bar is displayed during roof generation.}
}
\value{
A 3D mesh of the roof model.
}
\description{
This function generates a 3D roof model from a straight skeleton.
}
