% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_roof.R
\name{generate_roof_beveled}
\alias{generate_roof_beveled}
\title{Generate Offset Polygon}
\usage{
generate_roof_beveled(ss, bevel_offset, max_height = 1, offset_roof = 0)
}
\arguments{
\item{offset}{}
}
\description{
Generate Offset Polygon
}
