% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_maxima_nodes.R
\name{identify_maxima_nodes}
\alias{identify_maxima_nodes}
\title{Identify Local and Global Maxima Nodes}
\usage{
identify_maxima_nodes(straight_skeleton)
}
\arguments{
\item{straight_skeleton}{The straight skeleton object.}
}
\value{
Vector containting maxima ids
}
\description{
Identify Local and Global Maxima Nodes
}
\keyword{internal}
