% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_offset_links_nodes.R
\name{insert_polygon_links_nodes}
\alias{insert_polygon_links_nodes}
\title{Insert Polygon Links/Nodes}
\usage{
insert_polygon_links_nodes(ss, new_links)
}
\arguments{
\item{offset}{}
}
\description{
Insert Polygon Links/Nodes
}
\keyword{internal}
