% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_offset_polygon.R
\name{plot_offset_polygon}
\alias{plot_offset_polygon}
\title{Plot Offset Polygons}
\usage{
plot_offset_polygon(
  offset_polygons,
  skeleton = NULL,
  plot_original_polygon = TRUE,
  color = NULL,
  border = "dodgerblue",
  xlim = NULL,
  ylim = NULL,
  linewidth = 1
)
}
\arguments{
\item{offset_polygons}{Default `NULL`. A list of data frames, where each data frame represents an offset polygon. Each data frame should have 'x' and 'y' columns representing coordinates.}

\item{skeleton}{Default `NULL`. The straight skeleton used to generate the offset polygon. Include this if you want to adjust the plot
range to encompass the entire original polygon.}

\item{plot_original_polygon}{Default `TRUE`. If `skeleton` passed as well, plots the original polygon.}

\item{color}{Default `NULL`. A color or palette function to generate the fill palette for the polygons' interiors.}

\item{border}{Default `grDevices::heat.colors`.  A color or palette function to generate the color palette for the offset polygons' borders.}

\item{xlim}{Default `NULL`. The x-axis limits as a vector of two values (min, max). If `NULL`, it calculates the limits from the data.}

\item{ylim}{Default `NULL`. The y-axis limits as a vector of two values (min, max). If `NULL`, it calculates the limits from the data.}

\item{linewidth}{Default `1`. The linewidth of the polygon.}
}
\value{
A plot showing the offset polygons.
}
\description{
Plot the offset polygons generated by the `generate_offset_polygon` function.
}
\examples{
# Outer polygon
vertices = matrix(c(0,0, 7,0, 7,7, 0,7, 0,0), ncol = 2, byrow = TRUE)
# Holes inside the polygon
hole_1 = matrix(c(1,1, 2,1, 2,2, 1,2, 1,1), ncol = 2, byrow = TRUE)[5:1,]
hole_2 = matrix(c(5,5, 6,5, 6,6, 5,6, 5,5), ncol = 2, byrow = TRUE)[5:1,]
skeleton = skeletonize(vertices, holes = list(hole_1, hole_2))
plot_skeleton(skeleton)

#Generate three offsets
plot_offset_polygon(generate_offset_polygon(skeleton, c(0.25,0.75,1.5,2)))

#Generate three offsets, including the original polygon
plot_offset_polygon(generate_offset_polygon(skeleton, c(0.25,0.75,1.5,2)), skeleton = skeleton)

#Generate many offsets
plot_offset_polygon(generate_offset_polygon(skeleton, seq(0,2.5,by=0.1)))

# Skeletonize and plot an {sf} object
if(length(find.package("spData",quiet = TRUE)) > 0) {
  us_states = spData::us_states
  texas = us_states[us_states$NAME == "Texas",]
  texas_skeleton = skeletonize(texas)
  plot_offset_polygon(generate_offset_polygon(texas_skeleton, seq(0, 2.5, by = 0.1)),
                      border = heat.colors,
                      linewidth = 1)
}
}
