% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalculate_ordered_ids.R
\name{recalculate_ordered_ids}
\alias{recalculate_ordered_ids}
\title{Recalculate Ordered IDs}
\usage{
recalculate_ordered_ids(skeleton)
}
\arguments{
\item{ss}{}
}
\description{
Recalculate Ordered IDs
}
\keyword{internal}
