% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_reversed_links.R
\name{remove_reversed_links}
\alias{remove_reversed_links}
\title{Remove Reversed Links}
\usage{
remove_reversed_links(skeleton)
}
\arguments{
\item{ss}{}
}
\description{
Remove Reversed Links
}
\keyword{internal}
