% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletonize.R
\name{unit_vector}
\alias{unit_vector}
\title{Dot Product}
\usage{
unit_vector(v)
}
\arguments{
\item{v1}{Vector 1}

\item{v2}{Vector 2}
}
\value{
Value
}
\description{
Dot Product
}
\examples{
#None
}
\keyword{internal}
