% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_offset_polygon.R
\name{generate_offset_polygon}
\alias{generate_offset_polygon}
\title{Generate Offset Polygon}
\usage{
generate_offset_polygon(ss, offset)
}
\arguments{
\item{offset}{}
}
\description{
Generate Offset Polygon
}
