% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_roof.R
\name{generate_roof_beveled}
\alias{generate_roof_beveled}
\title{Generate a beveled 3D roof model}
\usage{
generate_roof_beveled(
  ss,
  bevel_offsets,
  bevel_heights = bevel_offsets,
  set_max_height = FALSE,
  max_height = 1,
  offset_roof = 0,
  bottom = FALSE,
  swap_yz = FALSE,
  progress = TRUE
)
}
\arguments{
\item{ss}{Default `NULL`. A straight skeleton generated from the `skeletonize` function.}

\item{bevel_offsets}{Default `NULL`. The offset(s) of the bevel.}

\item{bevel_heights}{Default is set to `bevel_offsets`. Numeric vector specifying the heights of the bevels. Must be of the same length as `bevel_offsets`.}

\item{set_max_height}{Default `FALSE`. A logical flag that controls whether to set the max height of the roof based on the `max_height` argument.}

\item{max_height}{Default `1`. The maximum height of the roof.}

\item{offset_roof}{Default `0`. The offset of the roof.}

\item{bottom}{Default `FALSE`. A logical flag that controls whether to generate the bottom of the roof.}

\item{swap_yz}{Default `FALSE`. A logical flag that controls whether to swap the y and z coordinates in the resulting mesh.
If `TRUE`, the y and z coordinates will be swapped.}

\item{progress}{Default `TRUE`. A logical flag to control whether a progress bar is displayed during roof generation.}
}
\value{
A 3D mesh of the beveled roof model.
}
\description{
This function generates a beveled 3D roof model from a straight skeleton.
}
