% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_roof.R
\name{generate_roof_beveled}
\alias{generate_roof_beveled}
\title{Generate a beveled 3D roof model}
\usage{
generate_roof_beveled(
  ss,
  bevel_offsets,
  bevel_heights = bevel_offsets,
  set_max_height = FALSE,
  max_height = 1,
  offset_roof = 0,
  base = FALSE,
  bottom = FALSE,
  swap_yz = FALSE
)
}
\arguments{
\item{ss}{Default `NULL`. A straight skeleton generated from the `skeletonize` function.}

\item{max_height}{Default `1`. The maximum height of the roof.}

\item{offset_roof}{Default `0`. The offset of the roof.}

\item{base}{Default `FALSE`. A logical flag that controls whether to generate the base of the roof.}

\item{bottom}{Default `FALSE`. A logical flag that controls whether to generate the bottom of the roof.}

\item{bevel_offset}{Default `NULL`. The offset of the bevel.}
}
\value{
A 3D mesh of the beveled roof model.
}
\description{
This function generates a beveled 3D roof model from a straight skeleton.
}
