% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-functions.R
\name{interpolate_location}
\alias{interpolate_location}
\title{Interpolate Location}
\usage{
interpolate_location(node_start, node_end, height_start, height_end, height)
}
\description{
Interpolate Location
}
\keyword{internal}
