% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_bevel_with_skeleton.R
\name{modify_bevel_with_skeleton}
\alias{modify_bevel_with_skeleton}
\title{Modify bevel data using straight skeleton}
\usage{
modify_bevel_with_skeleton(bevel_offsets, bevel_heights, straight_skeleton)
}
\arguments{
\item{straight_skeleton}{list. Straight skeleton data structure.}

\item{bevel_height}{numeric vector. The existing heights for the bevel.}

\item{bevel_offset}{numeric vector. The existing offsets for the bevel.}
}
\value{
list. Modified bevel_height and bevel_offset vectors.
}
\description{
Modify bevel data using straight skeleton
}
