% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_offset_polygon.R
\name{plot_offset_polygon}
\alias{plot_offset_polygon}
\title{Plot Offset Polygons}
\usage{
plot_offset_polygon(
  offset_polygons,
  skeleton = NULL,
  main = "Offset Polygons",
  plot_original_polygon = TRUE,
  col = NA,
  border = "blue",
  xlim = NULL,
  ylim = NULL,
  add = FALSE,
  palette_border = NULL,
  palette_col = NULL,
  linewidth = 1
)
}
\arguments{
\item{offset_polygons}{Default `NULL`. A list of data frames, where each data frame represents an offset polygon. Each data frame should have 'x' and 'y' columns representing coordinates.}

\item{skeleton}{Default `NULL`. The straight skeleton used to generate the offset polygon. Include this if you want to adjust the plot
range to encompass the entire original polygon.}

\item{main}{Default "Offset Polygons". The main title for the plot.}

\item{col}{Default "lightblue". The fill color for the polygons. This argument is ignored if `palette_col` is provided.}

\item{border}{Default "blue". The border color for the polygons. This argument is ignored if `palette_border` is provided.}

\item{xlim}{Default `NULL`. The x-axis limits as a vector of two values (min, max). If `NULL`, it calculates the limits from the data.}

\item{ylim}{Default `NULL`. The y-axis limits as a vector of two values (min, max). If `NULL`, it calculates the limits from the data.}

\item{add}{Default `FALSE`. A logical indicating whether to add the offset polygons to an existing plot (`TRUE`) or create a new one (`FALSE`).}

\item{palette_border}{Default `NULL`. A function to generate the color palette for the polygons' borders . This function should accept the number of polygons as input and return a vector of colors. If provided, it overrides the `border` argument.}

\item{palette_col}{Default `NULL`. A function to generate the color palette for the polygons' interiors. This function should accept the number of polygons as input and return a vector of colors. If provided, it overrides the `col` argument.}

\item{linewidth}{Default `1`. The linewidth of the polygon.}
}
\value{
A plot showing the offset polygons.
}
\description{
Plot the offset polygons generated by the `generate_offset_polygon` function.
}
\examples{
# Outer polygon
vertices = matrix(c(0,0, 7,0, 7,7, 0,7, 0,0), ncol = 2, byrow = TRUE)
# Holes inside the polygon
hole_1 = matrix(c(1,1, 2,1, 2,2, 1,2, 1,1), ncol = 2, byrow = TRUE)[5:1,]
hole_2 = matrix(c(5,5, 6,5, 6,6, 5,6, 5,5), ncol = 2, byrow = TRUE)[5:1,]
skeleton = skeletonize(vertices, holes = list(hole_1, hole_2))
plot_skeleton(skeleton)
plot_offset_polygon(generate_offset_polygon(skeleton, seq(0,2.5,by=0.1)), add = TRUE)
# Now with colors
plot_offset_polygon(generate_offset_polygon(skeleton, seq(0,2.5,by=0.1)),
                    palette_border = terrain.colors, skeleton = skeleton,
                    linewidth = 3)

# Skeletonize and plot an {sf} object
if(length(find.package("spData",quiet = TRUE)) > 0) {
  us_states = spData::us_states
  texas = us_states[us_states$NAME == "Texas",]
  texas_skeleton = skeletonize(texas)
  plot_offset_polygon(generate_offset_polygon(texas_skeleton, seq(0, 2.5, by = 0.1)),
                      palette_border = heat.colors, skeleton = texas_skeleton,
                      linewidth = 2)
}
}
