% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_skeleton.R
\name{plot_skeleton}
\alias{plot_skeleton}
\title{Plot the Straight Skeleton of a Polygon}
\usage{
plot_skeleton(skeleton)
}
\arguments{
\item{skeleton}{A list object of class 'rayskeleton' containing the straight
skeleton details. It should have 'nodes' and 'links' as its primary components.}
}
\value{
A plot visualizing the straight skeleton and the original polygon.
}
\description{
This function visualizes the straight skeleton derived from a given polygon.
The original polygon (with holes if present) is plotted in black, while the
straight skeleton is plotted in red.
}
\details{
The function uses base R plotting capabilities. The original graphics
         parameters (`par`) are restored after plotting to ensure no unintended
         side-effects.
}
\examples{
# Assuming skeleton1 is already defined as in the previous example
# Outer polygon
vertices = matrix(c(0,0, 7,0, 7,7, 0,7, 0,0), ncol = 2, byrow = TRUE)
# Holes inside the polygon
hole1 = matrix(c(1,1, 1,2, 2,2, 2,1, 1,1), ncol = 2, byrow = TRUE)
hole2 = matrix(c(5,5, 5,6, 6,6, 6,5, 5,5), ncol = 2, byrow = TRUE)
skeleton = skeletonize(vertices, holes = list(hole1, hole2))
plot_skeleton(skeleton)
}
