% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletonize.R
\name{skeletonize}
\alias{skeletonize}
\title{Title}
\usage{
skeletonize(
  vertices,
  holes = list(),
  tol = 1e-10,
  debug = FALSE,
  return_raw_ss = FALSE
)
}
\arguments{
\item{vertices}{DUM}

\item{holes}{DEE}

\item{tol}{DOO}

\item{debug}{}
}
\value{
Value
}
\description{
Title
}
\examples{
#Value
}
